flake.lock: update
• Updated input 'nixpkgs':
    'github:NixOS/nixpkgs/20075955deac2583bb12f07151c2df830ef346b4?narHash=sha256-HMwfAJBdrr8wXAkbGhtcby1zGFvs%2BStOp19xNsbqdOg%3D' (2025-08-19)
  → 'github:NixOS/nixpkgs/3b9f00d7a7bf68acd4c4abb9d43695afb04e03a5?narHash=sha256-XexyKZpf46cMiO5Vbj%2BdWSAXOnr285GHsMch8FBoHbc%3D' (2025-08-25)

+++ post-build-hook-queue/src/main.rs
@@ -36,10 +36,10 @@ fn run_timeout(child: io::Result<Child>, timeout: Duration) -> anyhow::Result<Op
             return Ok(status.code());
         } else if elapsed > timeout {
             log::error!("Timeout of {timeout:?} exceeded, killing child");
-            if let Err(e) = child.kill() {
-                if !matches!(e.kind(), ErrorKind::InvalidInput) {
-                    anyhow::bail!("Failed to kill child process: {e:?}")
-                }
+            if let Err(e) = child.kill()
+                && !matches!(e.kind(), ErrorKind::InvalidInput)
+            {
+                anyhow::bail!("Failed to kill child process: {e:?}")
             }
             return Ok(None);
         }
